Fixed positioning issues & syntax
I fixed the issues with the CSS. I commented out the parts causing the issues so you can review. I also added in a bleed and roomy class. We use these alot in our themes. Lastly, I updated the syntax in your html to my preference but I am not saying it was wrong before. Let me know if you have any questions